Operation.rb added and changes done
