+++ lib/adapter/mysqlConnect.rb
@@ -1,9 +1,10 @@
 require 'mysql'
-
+require_relative '../../config/initializer/config'
 module Mysqlapi
   def self.connect
+
     begin
-      con = Mysql.new('localhost', 'root', 'root', 'ruby')
+      con = Mysql.new(HOSTNAME, USERNAME, PASSWORD, DB_NAME)
       rs = con.query 'SELECT VERSION()'
     rescue Mysql::Error => e
       puts e.error
